refactor(Member, Favorite): Favorite이 회원을 간접 참조하게 수정
