+++ src/main/java/nextstep/member/ui/MemberControllerExceptionHandler.java
@@ -1,6 +1,5 @@
 package nextstep.member.ui;
 
-import nextstep.member.exception.FavoriteNotFoundException;
 import nextstep.member.exception.MemberAuthenticationException;
 import nextstep.member.exception.MemberNotFoundException;
 import nextstep.member.exception.TokenAuthorizationException;
@@ -25,11 +24,6 @@ public class MemberControllerExceptionHandler {
     @ExceptionHandler(TokenAuthorizationException.class)
     public ResponseEntity<Void> handleIllegalArgsException(TokenAuthorizationException e) {
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-    }
-
-    @ExceptionHandler(FavoriteNotFoundException.class)
-    public ResponseEntity<Void> handleIllegalArgsException(FavoriteNotFoundException e) {
-        return ResponseEntity.badRequest().build();
     }
 
 }
