+++ src/main/java/nextstep/subway/domain/FavoriteRepository.java
@@ -1,6 +1,9 @@
-package nextstep.member.domain;
+package nextstep.subway.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
+    List<Favorite> findByMemberId(Long id);
 }
